requirements.txt
python-dotenv
pandas
numpy
matplotlib
statsmodels
pandas-datareader
pandas_ta
datetime
yfinance
scikit-learn
pyportfolioopt
arch
Flask
Flask-CORS
xlsxwriter
lumibot
alpaca-trade-api
timedelta
streamlit
google-generativeai
langchain
PyPDF2
faiss-cpu
langchain_google_genai
chromadb
ipykernel